refactor: 레시피 등록페이지의 레시피 정보 영역 컴포넌트 분리(#208)
